cloud-function/requirements.txt
autopep8==2.0.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
gensim==4.2.0
google-api-core==2.10.2
google-auth==2.14.1
google-cloud-core==2.3.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
idna==3.4
joblib==1.2.0
nltk==3.7
numpy==1.23.5
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pymediawiki==0.7.2
regex==2022.10.31
requests==2.28.1
rsa==4.9
scipy==1.9.3
six==1.16.0
smart-open==6.2.0
soupsieve==2.3.2.post1
tomli==2.0.1
tqdm==4.64.1
urllib3==1.26.13